backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
httpx==0.25.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
# PayPal SDK
paypalrestsdk==1.13.1
# Google Auth
google-auth==2.23.3
# Apple Auth
pyjwt==2.8.0
cryptography==41.0.5
# Image handling
pillow==10.1.0 